isolation des sessions

+++ src/model/category_model.py
@@ -4,58 +4,54 @@ class Category:
     def __init__(self):
         self.conn = get_connection()
 
-    def create_category(self, name):
+    def create_category(self, name, user_id):
         cursor = self.conn.cursor()
         cursor.execute("""
-            INSERT INTO categories (name) VALUES (?)
-""",(name,))
+            INSERT INTO categories (name, user_id) VALUES (?, ?)
+""", (name, user_id))
         self.conn.commit()
         cursor.close()
 
-    def get_category_by_name(self, name):
+    def get_category_by_name(self, name, user_id):
         cursor = self.conn.cursor()
         cursor.execute("""
-            SELECT * FROM categories WHERE name = ?
-""", (name,)
-        )
+            SELECT * FROM categories WHERE name = ? AND user_id = ?
+""", (name, user_id))
         row = cursor.fetchone()
         cursor.close()
         return row
     
-    def get_category_by_id(self, category_id):
+    def get_category_by_id(self, category_id, user_id):
         cursor = self.conn.cursor()
         cursor.execute("""
-            SELECT * FROM categories WHERE id = ?
-""", (category_id,)
-        )
-        rows = cursor.fetchone()
+            SELECT * FROM categories WHERE id = ? AND user_id = ?
+""", (category_id, user_id))
+        row = cursor.fetchone()
         cursor.close()
-        return rows
+        return row
     
-    def get_all_category(self):
+    def get_all_category(self, user_id):
         cursor = self.conn.cursor()
         cursor.execute("""
-            SELECT * FROM categories ORDER BY NAME
-"""
-        )
+            SELECT * FROM categories WHERE user_id = ? ORDER BY NAME
+""", (user_id,))
         rows = cursor.fetchall()
         cursor.close()
         return rows
     
-    def update_category(self, category_id, new_name):
+    def update_category(self, category_id, new_name, user_id):
         cursor = self.conn.cursor()
         cursor.execute("""
-                UPDATE categories SET name = ? WHERE id = ?
-
-""", (new_name, category_id))
+                UPDATE categories SET name = ? WHERE id = ? AND user_id = ?
+""", (new_name, category_id, user_id))
         self.conn.commit()
         cursor.close()
 
-    def delete_category(self, category_id):
+    def delete_category(self, category_id, user_id):
         cursor = self.conn.cursor()
         cursor.execute("""
-            DELETE FROM categories WHERE id = ?
-""", (category_id,))
+            DELETE FROM categories WHERE id = ? AND user_id = ?
+""", (category_id, user_id))
         self.conn.commit()
         cursor.close()
